Encapsulate opentest4j classes in Failures.kt
